requirements.txt
typer>=0.9.0
rich>=13.0.0
tree-sitter>=0.21.0
tree-sitter-java>=0.21.0
Jinja2>=3.1.3
langchain>=0.1.16
langchain-google-genai>=1.0.3
pydantic>=2.7.0
python-dotenv>=1.0.0
PyYAML>=6.0
